put a preprocessor guard around MPI-4 large count stuff in WRAP
this way, WRAP function pointers will be null when they are not implemented, and we can detect in the outer layer.
there are more precise, nuanced ways to do this, but they are tedious and unnecessary, since it is reasonable
to expect that Open-MPI will implement large-count when they increment MPI_VERSION to 4.

Signed-off-by: Jeff Hammond <jeff.science@gmail.com>