split mode constant handling and fix values
modes need to be powers of two to allow OR-ing them.
split the conversion of IO and RMA but keep them in the
same enum since there might be some value in ensuring
they never conflict, in case some day somebody wants
to reuse them.

this fixes the error in IMB-IO during MPI_File_open.

Signed-off-by: Jeff Hammond <jeff.science@gmail.com>